Space after '%' in "unbind %;"
After swtich to the new command parser, '%' that is not followed by a
whitespace is now an error, as it is also used to start keywords, such
as '%if'.

It makes sense to surround ';' with spaces when it is used to separate
statements in any case - just to make sure it is not parsed as part of
an argument to the command on the left.